+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -2,14 +2,13 @@ package com.techcourse;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
+import com.interface21.webmvc.servlet.mvc.tobe.mapping.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
-import com.techcourse.adapter.AnnotationHandlerAdapter;
-import com.techcourse.adapter.HandlerAdapter;
-import com.techcourse.adapter.HandlerAdapterRegistry;
-import com.techcourse.adapter.ManualHandlerAdapter;
-import com.techcourse.mapping.HandlerMappingRegistry;
-import com.techcourse.mapping.ManualHandlerMapping;
+import com.interface21.webmvc.servlet.mvc.tobe.adapter.AnnotationHandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.tobe.adapter.HandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.tobe.adapter.HandlerAdapterRegistry;
+import com.interface21.webmvc.servlet.mvc.tobe.adapter.ManualHandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.tobe.mapping.HandlerMappingRegistry;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
