+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,4 +1,4 @@
-package com.techcourse.mapping;
+package com.techcourse;
 
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.asis.ForwardController;
